Duplicated file and methods deleted
